requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
django-filter~=23.1
psycopg2
pytest~=7.2.2